Added method to Tweeter that retrieves a User via its userId fixes TweetWallFX/TweetwallFX#428
